server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["ruihan00 <ruixhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.1"
pillow = "^10.4.0"
uvicorn = "^0.30.6"
pydantic = "^2.8.2"
python-socketio = "^5.11.3"
aiofiles = "^24.1.0"
websockets = "^12.0"
ultralytics = "^8.2.79"
torch = "^2.4.0"
torchvision = "^0.19.0"
opencv-python-headless = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
